classe per creazione workflow
