+++ frontend/src/services/newWorkflowService.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe, vi, beforeEach } from "vitest";
-import { newWorkflow } from "./newWorkflowAPI";
+import { NewWorkflowService } from "./newWorkflowService";
 import { API_BASE_URL } from "./constants";
 
 describe("newWorkflow API", () => {
@@ -9,17 +9,19 @@ describe("newWorkflow API", () => {
         fetchSpy.mockReset();
     });
 
+    const name = "testWorkflow";
+    const accessToken = "testToken";
+    const service = new NewWorkflowService(accessToken);
+
     test("Should return the created workflow when successful", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        const mockResponse = { name, accessToken };
+        const mockResponse = { name };
         
         fetchSpy.mockResolvedValue({
             status: 201,
             json: () => Promise.resolve(mockResponse)
         } as Response);
         
-        await expect(newWorkflow(name, accessToken)).resolves.toEqual(mockResponse);
+        await expect(service.newWorkflow(name)).resolves.toEqual(mockResponse);
         expect(fetchSpy).toBeCalledTimes(1);
         expect(fetchSpy).toBeCalledWith(`${API_BASE_URL}/workflow/create/${name}`, {
             method: "POST",
@@ -32,35 +34,26 @@ describe("newWorkflow API", () => {
     });
 
     test("Should throw an error with message 'Workflow with the same name already exists' if status 400 is received", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockResolvedValue({
             status: 400,
             json: () => Promise.resolve("Workflow with the same name already exists")
         } as Response);
         
-        await expect(newWorkflow(name, accessToken)).rejects.toThrowError("Workflow with the same name already exists");
+        await expect(service.newWorkflow(name)).rejects.toThrowError("Workflow with the same name already exists");
     });
 
     test("Should throw an error with message 'Server error' if status 500 is received", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockResolvedValue({
             status: 500,
             json: () => Promise.resolve("Server error")
         } as Response);
         
-        await expect(newWorkflow(name, accessToken)).rejects.toThrowError("Server error");
+        await expect(service.newWorkflow(name)).rejects.toThrowError("Server error");
     });
 
     test("Should throw an error with message 'Generic Error' if fetch fails", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockRejectedValue(new Error("Generic Error"));
         
-        await expect(newWorkflow(name, accessToken)).rejects.toThrowError("Generic Error");
+        await expect(service.newWorkflow(name)).rejects.toThrowError("Generic Error");
     });
 });
